pyproject.toml
[tool.poetry]
name = "colosseum_website"
version = "0.1.0"
description = "Website for colosseum"
authors = ["h3nnn4n <colosseum@h3nnn4n.me>"]
license = "MIT"

[tool.poetry.dependencies]
python = "3.10.0"
Django = "^3.2.9"
psycopg2 = "^2.9.2"
gunicorn = "^20.1.0"
dj-database-url = "^0.5.0"
django-heroku = "^0.3.1"
whitenoise = "^5.3.0"
django-extensions = "^3.1.5"
djangorestframework = "^3.12.4"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
